medical-claims-generator/requirements-dev.txt
# Development dependencies
-r requirements.txt

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
hypothesis>=6.90.0

# Code quality
black>=23.0.0
ruff>=0.1.0
mypy>=1.5.0

# Type stubs
types-reportlab>=4.0.0
pandas-stubs>=2.0.0
boto3-stubs[s3]>=1.28.0
